requirements.txt
accelerate==0.25.0
albucore==0.0.15
albumentations==1.4.15
apex==0.1
blobfile==2.1.1
datasets==3.0.0
deepspeed==0.6.5
diffusers==0.28.2
dotmap==1.3.30
easydict==1.13
einops==0.7.0
ffmpeg==1.4
fvcore 
glob2==0.7
h5py==3.10.0
hjson==3.1.0
huggingface-hub==0.24.7
imageio==2.33.1
imageio-ffmpeg==0.4.9
insightface==0.7.3
iopath==0.1.9
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
json5==0.9.5
jsonschema==3.2.0
kornia==0.7.3
kornia-rs==0.1.5
matplotlib==3.8.2
moviepy==1.0.3
munch==4.0.0
ninja==1.11.1.1
numpy==1.26.2
onnx==1.16.2
onnxruntime==1.16.3
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.84
pandas==2.1.4
peft==0.12.0
Pillow==10.1.0
protobuf==3.20.2
pytorch-msssim==1.0.0
pytorch3d==0.7.5
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy
sk-video==1.1.10
timm==0.5.4
tokenizers==0.19.1
torch==2.0.1
torchaudio==2.0.2
torchinfo==1.8.0
torchvision==0.15.2
tqdm==4.66.5
transformers==4.44.2
trimesh==4.0.8
triton==2.0.0
typing-extensions==4.12.2
wandb==0.16.1
xformers==0.0.22.post7
facenet-pytorch==2.5.3
youtube_dl